+++ src/components/UI/docs/index.ts
@@ -3,6 +3,6 @@ export * from './Code';
 export * from './DocsLinks';
 export * from './DocsNav';
 export * from './DocumentNav';
-export * from './LinkList';
+export * from './LinksList';
 export * from './Note';
 export { default } from './MDComponents';
